refactor(provision): move windows local logic into dedicated file
